requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
openai==1.6.1
chromadb==0.4.22
sentence-transformers==2.2.2
tiktoken==0.5.2
Pillow==10.1.0
email-validator==2.1.0
